Add requirements.txt to specify project dependencies.
Cleaned up code, removed depreciated methods. Originally had intended to allow through the UI an ability to pull your own data. However chose to remove this as there was a change in OSM. Vienna's city structure no longer is boundries, meaning that you can't search for 1180 Wien, and get the district as a polygon and pull all restuarants in the area.

Data in csv still works.